stmt

+++ code_gen2/lib/dyn.cc
@@ -49,14 +49,14 @@ dyn list(const vector<dyn>& v) {
 	return dyn(r, t_list);
 }
 
-dyn list(int a, dyn b) {
+dyn list(size_t a, dyn b) {
 	auto r = list(2);
 	r->v[0] = sym(a);
 	r->v[1] = b;
 	return dyn(r, t_list);
 }
 
-dyn list(int a, dyn b, dyn c) {
+dyn list(size_t a, dyn b, dyn c) {
 	auto r = list(3);
 	r->v[0] = sym(a);
 	r->v[1] = b;
@@ -64,7 +64,7 @@ dyn list(int a, dyn b, dyn c) {
 	return dyn(r, t_list);
 }
 
-dyn list(int a, dyn b, dyn c, dyn d) {
+dyn list(size_t a, dyn b, dyn c, dyn d) {
 	auto r = list(4);
 	r->v[0] = sym(a);
 	r->v[1] = b;
